+++ src/gamspy/_model.py
@@ -168,10 +168,10 @@ class Model:
     equations : str | Iterable
         List of Equation objects or str. ``all`` as a string represents
         all the equations specified before the creation of this model
-    problem : str
-        Problem type (e.g. LP, NLP etc.)
-    sense : "MIN", "MAX", or "FEASIBILITY", optional
-        Minimize or maximize
+    problem : Problem
+        'LP', 'NLP', 'QCP', 'DNLP', 'MIP', 'RMIP', 'MINLP', 'RMINLP', 'MIQCP', 'RMIQCP', 'MCP', 'CNS', 'MPEC', 'RMPEC', 'EMP', or 'MPSGE'
+    sense : Sense, optional
+        "MIN", "MAX", or "FEASIBILITY"
     objective : Variable | Expression, optional
         Objective variable to minimize or maximize or objective itself
     limited_variables : Iterable, optional
@@ -194,9 +194,9 @@ class Model:
         self,
         container: Container,
         name: str,
-        problem: str,
-        equations: tuple[Equation] | list[Equation] = [],
-        sense: Literal["MIN", "MAX", "FEASIBILITY"] | None = None,
+        problem: Problem,
+        equations: list[Equation] = [],
+        sense: Sense | None = None,
         objective: Variable | Expression | None = None,
         matches: dict | None = None,
         limited_variables: Iterable[Variable] | None = None,
@@ -252,6 +252,15 @@ class Model:
             f"{self._generate_prefix}{self.name}_{attr_name}"
             for attr_name in attribute_map.keys()
         ]
+
+    def __repr__(self) -> str:
+        return f"<Model `{self.name}` ({hex(id(self))})>"
+
+    def __str__(self) -> str:
+        return (
+            f"Model {self.name}:\n  Problem Type: {self.problem}\n  Sense:"
+            f" {self.sense}\n  Equations: {self.equations}"
+        )
 
     def _set_objective_variable(
         self,